chore: code refactoring (see desc)
- move sign to shared once it has been implemented for v3
- removed console.log
