+++ src/v2/sign/sign.test.ts
@@ -1,6 +1,6 @@
 import { sign as EcdsaSecp256k1Signature2019 } from "./ecdsa-secp256k1-signature-2019";
 import { ethers } from "ethers";
-import sampleDoc from "../../v2/schema/sample-document.json";
+import sampleDoc from "../schema/sample-document.json";
 import { ProofType } from "../../shared/@types/document";
 import { wrapDocument } from "../..";
